Revert "Split ATen/Parallel into interface and backend (#20057)", "Use AT_INTERNAL_ASSERT in test_base (#20555)"
This reverts commit c4a3b4d528db2e2a5c1c466e61fb740ee6ce0b03.
This reverts commit 70eb315da441701e0026e2127c466cf8eef4d005.
