+++ aten/src/ATen/Parallel.h
@@ -1,5 +1,16 @@
 #pragma once
 #include <ATen/ATen.h>
+#include <c10/core/thread_pool.h>
+
+#include <atomic>
+#include <cstddef>
+#include <exception>
+
+#ifdef _OPENMP
+#define INTRA_OP_PARALLEL
+
+#include <omp.h>
+#endif
 
 namespace at {
 namespace internal {
@@ -26,17 +37,56 @@ CAFFE2_API int get_num_threads();
 
 // Returns the current thread number (starting from 0)
 // in the current parallel region, or 0 in the sequential region
-CAFFE2_API int get_thread_num();
+inline int get_thread_num() {
+#ifdef _OPENMP
+  return omp_get_thread_num();
+#else
+  return 0;
+#endif
+}
 
-// Checks whether the code runs in parallel region
-CAFFE2_API bool in_parallel_region();
+inline bool in_parallel_region() {
+#ifdef _OPENMP
+  return omp_in_parallel();
+#else
+  return false;
+#endif
+}
 
 template <class F>
 inline void parallel_for(
     const int64_t begin,
     const int64_t end,
     const int64_t grain_size,
-    const F& f);
+    const F& f) {
+#ifdef _OPENMP
+  std::atomic_flag err_flag = ATOMIC_FLAG_INIT;
+  std::exception_ptr eptr;
+#pragma omp parallel if (!omp_in_parallel() && ((end - begin) >= grain_size))
+  {
+    int64_t num_threads = omp_get_num_threads();
+    int64_t tid = omp_get_thread_num();
+    int64_t chunk_size = divup((end - begin), num_threads);
+    int64_t begin_tid = begin + tid * chunk_size;
+    if (begin_tid < end) {
+      try {
+        f(begin_tid, std::min(end, chunk_size + begin_tid));
+      } catch (...) {
+        if (!err_flag.test_and_set()) {
+          eptr = std::current_exception();
+        }
+      }
+    }
+  }
+  if (eptr) {
+    std::rethrow_exception(eptr);
+  }
+#else
+  if (begin < end) {
+    f(begin, end);
+  }
+#endif
+}
 
 /*
 parallel_reduce
@@ -77,11 +127,37 @@ inline scalar_t parallel_reduce(
     const int64_t end,
     const int64_t grain_size,
     const scalar_t ident,
-    const F& f,
-    const SF& sf);
+    const F f,
+    const SF sf) {
+  if (in_parallel_region() || get_num_threads() == 1) {
+    return f(begin, end, ident);
+  } else {
+    const int64_t num_results = divup((end - begin), grain_size);
+    std::vector<scalar_t> results(num_results);
+    scalar_t* results_data = results.data();
+#ifdef _OPENMP
+#pragma omp parallel for if ((end - begin) >= grain_size)
+#endif
+    for (int64_t id = 0; id < num_results; id++) {
+      int64_t i = begin + id * grain_size;
+      results_data[id] = f(i, i + std::min(end - i, grain_size), ident);
+    }
+    return std::accumulate(
+        results_data, results_data + results.size(), ident, sf);
+  }
+}
 
 // Returns a detailed string describing parallelization settings
 CAFFE2_API std::string get_parallel_info();
+
+class CAFFE2_API PTThreadPool : public c10::ThreadPool {
+ public:
+  explicit PTThreadPool(
+      int pool_size,
+      int numa_node_id = -1);
+
+  void init_thread() override;
+};
 
 // Sets number of threads used for inter-op parallelism
 CAFFE2_API void set_num_interop_threads(int);
@@ -93,7 +169,3 @@ CAFFE2_API int get_num_interop_threads();
 CAFFE2_API void launch(const std::function<void()>& func);
 
 } // namespace at
-
-#if AT_PARALLEL_OPENMP
-#include <ATen/ParallelOpenMP.h>
-#endif
