+++ caffe2/CMakeLists.txt
@@ -862,16 +862,6 @@ torch_set_target_props(caffe2)
 #endif()
 
 target_compile_options(caffe2 PRIVATE "-DCAFFE2_BUILD_MAIN_LIB")
-
-# Parallelism settings
-#  OPENMP - OpenMP for intra-op, native thread pool for inter-op parallelism
-set(PARALLEL_BACKEND "OPENMP" CACHE STRING "ATen parallel backend")
-if ("${PARALLEL_BACKEND}" STREQUAL "OPENMP")
-  target_compile_definitions(caffe2 PUBLIC "-DAT_PARALLEL_OPENMP=1")
-else()
-  message(FATAL_ERROR "Unknown parallel backend: ${PARALLEL_BACKEND}")
-endif()
-
 if (MSVC AND NOT BUILD_SHARED_LIBS)
   # Note [Supporting both static and dynamic libraries on Windows]
   # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
