+++ setup.py
@@ -149,11 +149,6 @@
 #   LIBRARY_PATH
 #   LD_LIBRARY_PATH
 #     we will search for libraries in these paths
-#
-#   PARALLEL_BACKEND
-#     parallel backend to use for intra- and inter-op parallelism
-#     possible values:
-#       OPENMP - use OpenMP for intra-op and native backend for inter-op tasks
 
 from __future__ import print_function
 from setuptools import setup, Extension, distutils, find_packages
