+++ tools/build_pytorch_libs.py
@@ -226,10 +226,6 @@ def run_cmake(version,
     if mkldnn_threading:
         cmake_defines(cmake_args, MKLDNN_THREADING=mkldnn_threading)
 
-    parallel_backend = os.getenv('PARALLEL_BACKEND')
-    if parallel_backend:
-        cmake_defines(cmake_args, PARALLEL_BACKEND=parallel_backend)
-
     if USE_GLOO_IBVERBS:
         cmake_defines(cmake_args, USE_IBVERBS="1", USE_GLOO_IBVERBS="1")
